+++ lib/paperclip-storage-tmp/testing/cucumber.rb
@@ -1,0 +1,3 @@
+require 'paperclip-storage-tmp'
+
+After { Paperclip::Storage::Tmp.clear }