+++ lib/paperclip-storage-tmp/testing/rspec.rb
@@ -1,0 +1,5 @@
+require 'paperclip-storage-tmp'
+
+RSpec.configure do |config|
+  config.after { Paperclip::Storage::Tmp.clear }
+end