+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Not released yet:
+
+* Provide one-line testing helpers for RSpec and Cucumber;
+  they require the gem and add after hooks which clear the
+  tmp storage
+
 ## 0.0.1 (2012-05-11):
 
 * Implement `Paperclip::Storage::Tmp.clear` which clears all